# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DCRM API",
    "description": "API REST du CRM interne DCB Technologies. Format de réponse unifié : succès `{ success: true, data, meta? }`, erreur `{ success: false, error: { code, message } }`. Authentification par JWT Bearer (15 min, refresh 7 j via cookie httpOnly) ou par clé API longue durée (header X-API-Key). L'extension `x-permission` sur une opération indique la clé de permission RBAC requise. Tous les montants monétaires sont exprimés en euros hors taxes (HT) ; le TTC n'est jamais stocké et se dérive via Product.vatRate (défaut 20 %).",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Développement"
    },
    {
      "url": "https://{domaine}/api",
      "description": "Production",
      "variables": {
        "domaine": {
          "default": "crm.example.fr"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token obtenu via POST /auth/login (durée 15 min)"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Clé API longue durée générée via POST /apikeys (préfixe dcrm_). Permissions effectives = intersection des droits de la clé (scopes) et des droits actuels de son propriétaire — jamais de bypass '*', même pour un ADMIN ; les routes gardées par rôle sont inaccessibles par clé API."
      }
    },
...
  },
  "paths": {
...
    "/equipment": {
      "get": {
        "tags": [
          "Équipements"
        ],
        "summary": "Lister les équipements (paginé, filtrable)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "equipment:read",
        "parameters": [
          {
            "name": "companyId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warrantyExpiringSoon",
            "in": "query",
            "description": "\"true\" → garantie expirant dans les 90 jours",
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée des équipements"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission equipment:read manquante"
          }
        }
      },
      "post": {
        "tags": [
          "Équipements"
        ],
        "summary": "Créer un équipement",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "equipment:create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "companyId",
                  "type"
                ],
                "properties": {
                  "companyId": {
                    "type": "string"
                  },
                  "contractId": {
                    "type": "string"
                  },
                  "productId": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "brand": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "serialNumber": {
                    "type": "string"
                  },
                  "purchaseDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "warrantyExpiry": {
                    "type": "string",
                    "format": "date"
                  },
                  "location": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Équipement créé"
          },
          "400": {
            "description": "VALIDATION_ERROR ; COMPANY_NOT_FOUND, CONTRACT_NOT_FOUND, PRODUCT_NOT_FOUND si l'id fourni est introuvable ; CONTRACT_COMPANY_MISMATCH si le contrat lié appartient à une autre société"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission equipment:create manquante"
          }
        }
      }
    },
    "/equipment/{id}": {
      "get": {
        "tags": [
          "Équipements"
        ],
        "summary": "Détail d'un équipement",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "equipment:read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Équipement (company, contract, 10 derniers tickets, licenses)"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission equipment:read manquante"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      },
      "put": {
        "tags": [
          "Équipements"
        ],
        "summary": "Mettre à jour un équipement",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "equipment:update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companyId": {
                    "type": "string"
                  },
                  "contractId": {
                    "type": "string"
                  },
                  "productId": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "brand": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "serialNumber": {
                    "type": "string"
                  },
                  "purchaseDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "warrantyExpiry": {
                    "type": "string",
                    "format": "date"
                  },
                  "location": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Équipement mis à jour"
          },
          "400": {
            "description": "VALIDATION_ERROR ; mêmes erreurs de cohérence que la création (COMPANY_NOT_FOUND, CONTRACT_NOT_FOUND, PRODUCT_NOT_FOUND, CONTRACT_COMPANY_MISMATCH), vérifiées sur la valeur effective"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission equipment:update manquante"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      },
      "delete": {
        "tags": [
          "Équipements"
        ],
        "summary": "Supprimer un équipement",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "equipment:delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Équipement supprimé"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission equipment:delete manquante"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      }
    },
...
  }
}